Refine i18n
