+++ console/src/locales/zh-CN.yaml
@@ -1034,19 +1034,19 @@ core:
       submit:
         button: 初始化
         toast_success: 初始化成功
+      setup_initial_data:
+        loading: 正在初始化数据，请稍后...
     fields:
       site_title:
-        placeholder: 站点名称
-        validation: 请输入站点名称
+        label: 站点名称
       email:
-        placeholder: 邮箱
-        validation: 请输入邮箱
+        label: 邮箱
       username:
-        placeholder: 用户名
-        validation: 请输入用户名
+        label: 用户名
       password:
-        placeholder: 密码
-        validation: 请输入密码
+        label: 密码
+      confirm_password:
+        label: 确认密码
   rbac:
     Attachments Management: 附件
     Attachment Manage: 附件管理
